fix: use stable resource names for connector jobs
The kubernetes connector templates for the create and delete jobs should
use stable names. The `randNumeric` should not be necessary to ensure a
new k8s job can be created the next time the helm chart is
upgraded/installed. Instead, the existing [helm deletion policy
annotation](https://helm.sh/docs/topics/charts_hooks/#hook-deletion-policies)
should be sufficient. The existing annotation ensures helm deletes the
k8s job after it succeeds and before helm creates a new version of the
job.

The random value in the name was causing issues with CD tooling that
runs an install/upgrade on every sync such as ArgoCD or Tanka. By using
a stable name, tooling will be aware the job was already executed. This
PR should resolve #137