Allow to run the broker wrapper script in verbose mode
When we run the broker on our Docker image, we have a wrapper script called:
/entrypoint.sh

That script parses various things from the environment and the filesystem
prior to launching the broker.

This script is set to run in "errxit mode" with:
set -e

This causes the script to silently fail without emitting its execution commands
and rendering `kubectl logs` empty when things go south (e.g. null TunnelClientAllowedDomains).

This commit allows deploying the Helm chart with verbosity, such that we can run:
$ kubectl logs wiz-kubernetes-connector-broker-84cb965b9b-6n4qc -n wiz

and trace its execution to the point it has failed during startup.

The newly added field is `global.broker.logLevel`, which by default is set to 'info'.

To enable it, set the `global.broker.logLevel=debug` in the values, e.g.
$ helm install wiz-eks-connector-test \
  ./wiz-kubernetes-connector/ \
  --values values_wiz-kubernetes-connector.yaml \
  --set global.broker.logLevel=debug \
  -n wiz \
  --create-namespace